Adds a RuboCop todo config
With the main config file (`.rubocop.yml`) inheriting from this new
file, we won't have any issues reported when we run RuboCop. The benefit
here is that any new issues will be reported.

I'm planning on addressing the todo's in this file before I start
anything else. As a team, we'll review the cops listed here. We'll
choose to either correct what is reported, or decide on a different
standard. In latter case, we'll add that standard to the main config
file, i.e. turn off the cop, or choose a different setting for it. For
example, I'm guessing for the Style/Documentation cop, we'll just
disable it in `.rubocop.yml`.